Roles
 - Added Spring security
 - Tested if we can check for roles
