+++ src/main/java/com/medheroes/app/AppApplication.java
@@ -2,48 +2,12 @@ package com.medheroes.app;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
-import org.springframework.security.crypto.factory.PasswordEncoderFactories;
-import org.springframework.security.crypto.password.PasswordEncoder;
 
 @SpringBootApplication
 public class AppApplication {
 
-	public static void main(String[] args) {
-		SpringApplication.run(AppApplication.class, args);
-	}
+    public static void main(String[] args) {
+        SpringApplication.run(AppApplication.class, args);
+    }
 
-	@Configuration
-	@EnableWebSecurity
-	public class BasicConfiguration extends WebSecurityConfigurerAdapter {
-
-		@Override
-		protected void configure(AuthenticationManagerBuilder auth) throws Exception {
-			PasswordEncoder encoder =
-					PasswordEncoderFactories.createDelegatingPasswordEncoder();
-			auth
-					.inMemoryAuthentication()
-					.withUser("user")
-					.password(encoder.encode("user"))
-					.roles("USER")
-					.and()
-					.withUser("admin")
-					.password(encoder.encode("admin"))
-					.roles("USER", "ADMIN");
-		}
-
-		@Override
-		protected void configure(HttpSecurity http) throws Exception {
-			http
-					.authorizeRequests()
-					.anyRequest()
-					.authenticated()
-					.and()
-					.formLogin();
-		}
-	}
 }
