--- src/main/resources/application.yml
@@ -1,5 +1,0 @@
-spring:
-  security:
-    user:
-      name: user1
-      password: user1
